Inline template package
