+++ core/src/main/java/com/graphhopper/routing/CHPathCalculator.java
@@ -16,10 +16,8 @@
  *  limitations under the License.
  */
 
-package com.graphhopper.routing.template;
-// todonow: rename this package?
+package com.graphhopper.routing;
 
-import com.graphhopper.routing.*;
 import com.graphhopper.routing.querygraph.QueryGraph;
 import com.graphhopper.storage.Graph;
 import com.graphhopper.util.Parameters;
